[BOT] sandeep/refactor: :fire: remove binary-utils package (#17169)
* refactor: :fire: remove binary-utils package

* refactor: :fire: moved common function into binary-utils.js
