+++ packages/bot-skeleton/package.json
@@ -35,13 +35,12 @@
         "typescript": "^4.6.3"
     },
     "dependencies": {
-        "blockly": "^10.4.3",
         "@deriv/deriv-api": "^1.0.15",
         "@deriv/indicators": "^1.0.0",
         "@deriv/js-interpreter": "^3.0.0",
         "@deriv/shared": "^1.0.0",
         "@deriv/translations": "^1.0.0",
-        "binary-utils": "^4.23.0",
+        "blockly": "^10.4.3",
         "file-saver": "^2.0.2",
         "immutable": "^3.8.2",
         "localforage": "^1.9.0",
