+++ packages/bot-skeleton/src/services/api/ticks_service.js
@@ -1,43 +1,16 @@
 import { Map } from 'immutable';
-import { historyToTicks, getLast } from 'binary-utils';
 import { doUntilDone, getUUID } from '../tradeEngine/utils/helpers';
 import { observer as globalObserver } from '../../utils/observer';
 import { api_base } from './api-base';
-
-const parseTick = tick => ({
-    epoch: +tick.epoch,
-    quote: +tick.quote,
-});
-
-const parseOhlc = ohlc => ({
-    open: +ohlc.open,
-    high: +ohlc.high,
-    low: +ohlc.low,
-    close: +ohlc.close,
-    epoch: +(ohlc.open_time || ohlc.epoch),
-});
-
-const parseCandles = candles => candles.map(t => parseOhlc(t));
-
-const updateTicks = (ticks, newTick) => (getLast(ticks).epoch >= newTick.epoch ? ticks : [...ticks.slice(1), newTick]);
-
-const updateCandles = (candles, ohlc) => {
-    const lastCandle = getLast(candles);
-    if (
-        (lastCandle.open === ohlc.open &&
-            lastCandle.high === ohlc.high &&
-            lastCandle.low === ohlc.low &&
-            lastCandle.close === ohlc.close &&
-            lastCandle.epoch === ohlc.epoch) ||
-        lastCandle.epoch > ohlc.epoch
-    ) {
-        return candles;
-    }
-    const prevCandles = lastCandle.epoch === ohlc.epoch ? candles.slice(0, -1) : candles.slice(1);
-    return [...prevCandles, ohlc];
-};
-
-const getType = isCandle => (isCandle ? 'candles' : 'ticks');
+import {
+    historyToTicks,
+    parseTick,
+    parseOhlc,
+    parseCandles,
+    updateTicks,
+    updateCandles,
+    getType,
+} from './binary-utils';
 
 export default class TicksService {
     constructor() {
